pyproject.toml
[tool.poetry]
name = "caloculator"
version = "0.1.0"
description = ""
authors = ["qu3vipon <sdh5813@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
djangorestframework = "<3.11.0"
boto3 = "^1.14.8"
awscli = "^1.18.85"
django-storages = "^1.9.1"
psycopg2-binary = "^2.8.5"
gunicorn = "^20.0.4"
supervisor = "^4.2.0"
django-phonenumber-field = "^4.0.0"
phonenumbers = "^8.12.6"
phonenumberslite = "^8.12.6"
django-rest-auth = "^0.9.5"
django-allauth = "^0.42.0"
pillow = "^7.1.2"
djangorestframework-simplejwt = "<4.4"

[tool.poetry.dev-dependencies]
django-extensions = "^2.2.9"
youtube-dl = "^2020.6.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
